@@ -1,11 +1,13 @@
 <Project>
   <PropertyGroup>
-    <Version>7.0.0-rc01</Version>
-    <LangVersion>11.0</LangVersion>
+    <Version>8.0.0-rc00</Version>
+    <LangVersion>12.0</LangVersion>
     <Nullable>enable</Nullable>
     <FxNugetVersion>6.0.0</FxNugetVersion>
     <Fx7NugetVersion>7.0.0</Fx7NugetVersion>
     <Fx3NugetVersion>3.1.16</Fx3NugetVersion>
+    <Fx8NugetVersion>8.0.0</Fx8NugetVersion>
     <IsTrimmable>true</IsTrimmable>
+    <IsAotCompatible>true</IsAotCompatible>
   </PropertyGroup>
 </Project>
