@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net7.0;net6.0;netstandard2.1</TargetFrameworks>
+    <TargetFrameworks>net8.0;net7.0;net6.0;netstandard2.1</TargetFrameworks>
     <RootNamespace>NCoreUtils.Images</RootNamespace>
     <!-- NuGet -->
     <PackageId>NCoreUtils.Images.Abstractions</PackageId>
@@ -26,8 +26,8 @@
     <Compile Remove="CompilerServices\**" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="NCoreUtils.Extensions.Memory" Version="7.0.0" />
-    <PackageReference Include="NCoreUtils.IO.Abstractions" Version="7.0.0" />
-    <PackageReference Include="NCoreUtils.Resources.Abstractions" Version="7.0.0" />
+    <PackageReference Include="NCoreUtils.Extensions.Memory" Version="8.0.0" />
+    <PackageReference Include="NCoreUtils.IO.Abstractions" Version="8.0.0" />
+    <PackageReference Include="NCoreUtils.Resources.Abstractions" Version="8.0.0" />
   </ItemGroup>
 </Project>
