@@ -2,12 +2,11 @@ using System.Runtime.CompilerServices;
 using ImageMagick;
 using NCoreUtils.Images.Internal;
 
-namespace NCoreUtils.Images.ImageMagick
+namespace NCoreUtils.Images.ImageMagick;
+
+internal static class RectangleExtensions
 {
-    internal static class RectangleExtensions
-    {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static MagickGeometry ToMagickGeometry(this in Rectangle rect)
-            => new(rect.X, rect.Y, rect.Width, rect.Height);
-    }
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static MagickGeometry ToMagickGeometry(this in Rectangle rect)
+        => new(rect.X, rect.Y, rect.Width, rect.Height);
 }
