@@ -2,9 +2,7 @@ using System.Numerics;
 
 namespace NCoreUtils.Images.ImageMagick;
 
-internal partial record struct BucketUsage(BigInteger Total, BigInteger Used);
-
-internal partial record struct BucketUsage
+internal readonly partial record struct BucketUsage(BigInteger Total, BigInteger Used)
 {
     public UsageData Snapshot() => new(Total, Used);
 }
